splitting alt header and variant queries and test to separate files
