+++ src/main/kotlin/net/maizegenetics/phgv2/utils/TIledbCoreHvcfUtils.kt
@@ -392,371 +392,6 @@ fun writeVariantsDataToTileDB(variantArrayName:String, combinedHvcfVariantData:L
     context.close()
 }
 
-/**
- * This function queries the TileDB array for the sampleName and Id associated with refernece ranges
- * in the array name provided.  THis was written assuming the query was for the alt_headers_array -
- * but the general principle should be adapted to both arrays.
- * It uses a stream to continually read results while the query statue is INCOMPLETE.
- * The function returns a list of maps, where each map contains the fields SampleName and ID.
- *
- * Streaming works with smallseq, so probably really isn't streaming.  Needs testing with
- * a larger dataset.
- *
- * This was tested using the alt_header_array, not the hvcf_variants_array
- * THis works when I grab ALL data, then filter at the end.
- */
-
-fun queryWithStreaming_sampleNameIdByRefRange(arrayName: String, refRangeList: List<String>): Map<String,List<Map<String,String>>> {
-    val context = Context()
-    //val combinedResult = mutableListOf<Map<String, String>>()
-    val resultMap = mutableMapOf<String, List<Map<String,String>>>()
-
-    // Open the array in read mode
-    val array = Array(context, arrayName, QueryType.TILEDB_READ)
-
-    // Prepare buffers for RefRange (filtering dimension)
-    val refRangeBuffers = prepareVariableBuffer(context, refRangeList)
-    println("queryWithStreaming: refRangeList size = ${refRangeList.size},  refRangeBuffers.first size = ${refRangeBuffers.first.size}, second size = ${refRangeBuffers.second.size}")
-    // print the refRangeBuffers.second as a long array
-    val refRangeOffsetsArray = refRangeBuffers.second.toJavaArray() as LongArray
-    println("queryWithStreaming: refRangeOffsetsArray = ${refRangeOffsetsArray.contentToString()}")
-
-    // Set reasonable initial buffer sizes
-    val sampleNameBufferSize = 8192
-    val sampleNameOffsetsSize = 1024
-    val idBufferSize = 8192
-    val idOffsetsSize = 1024
-    val refRangeBufferSize = 8192
-    val refRangeOffsetsSize = 1024
-
-    // Allocate initial buffers
-    val sampleNameBuffer = io.tiledb.java.api.NativeArray(context, sampleNameBufferSize, Datatype.TILEDB_STRING_ASCII)
-    val sampleNameOffsets = io.tiledb.java.api.NativeArray(context, sampleNameOffsetsSize, Datatype.TILEDB_UINT64)
-    val idBuffer = io.tiledb.java.api.NativeArray(context, idBufferSize, Datatype.TILEDB_STRING_ASCII)
-    val idOffsets = io.tiledb.java.api.NativeArray(context, idOffsetsSize, Datatype.TILEDB_UINT64)
-    val refRangeBuffer = io.tiledb.java.api.NativeArray(context, refRangeBufferSize, Datatype.TILEDB_STRING_ASCII)
-    val refRangeOffsets = io.tiledb.java.api.NativeArray(context, refRangeOffsetsSize, Datatype.TILEDB_UINT64)
-
-    // Create the query
-    val query = Query(array, QueryType.TILEDB_READ).apply {
-        setLayout(Layout.TILEDB_UNORDERED) // GLOBAL_ORDER didn't get all the data, and associated it incorrectly
-
-        // Set RefRange buffers
-//        setDataBuffer("RefRange", refRangeBuffer)
-//        setOffsetsBuffer("RefRange", refRangeOffsets)
-
-        // Set RefRange Buffers  - these are the buffers from prepareVariableBuffer(refRangeList)
-
-
-
-        setDataBuffer("RefRange", refRangeBuffers.first)
-        setOffsetsBuffer("RefRange", refRangeBuffers.second)
-
-        // Set SampleName buffers
-        setDataBuffer("SampleName", sampleNameBuffer)
-        setOffsetsBuffer("SampleName", sampleNameOffsets)
-
-        // Set ID buffers
-        setDataBuffer("ID", idBuffer)
-        setOffsetsBuffer("ID", idOffsets)
-    }
-
-    // Add ranges for RefRange dimension. This will allow tiledb to do the "slicing" for us
-    // This doesn't work due to conda issues.  It seems we need a core tiledb of 12.2.0 or later
-    // I tried adding and creating a new environment but could not get the env to build on the iMac.
-    // SO for now, we will just get all the data and filter it at the end.
-//    refRangeList.forEach { refRange ->
-//        query.addRange(0, refRange, refRange, null) // Constrain to specific RefRange values
-//    }
-
-    // Iterate until query is complete
-    do {
-        query.submit()
-
-        // Use resultBufferElements to determine the number of valid entries
-        val numSampleNames = query.resultBufferElements()["SampleName"]?.first?.toInt() ?: 0
-        val numIDs = query.resultBufferElements()["ID"]?.first?.toInt() ?: 0
-        val numRefRanges = query.resultBufferElements()["RefRange"]?.first?.toInt() ?: 0
-        println("queryWithStreaming the query.submit results show : num SampleNames=$numSampleNames, num IDs=$numIDs, num RefRanges=$numRefRanges")
-
-        // Extract RefRange data
-//        val refRangeOffsetsArray = refRangeOffsets.toJavaArray() as LongArray
-//        val refRangeRawData = String(refRangeBuffer.toJavaArray() as ByteArray)
-//        val refRanges = refRangeOffsetsArray.take(numRefRanges).mapIndexedNotNull { index, offset ->
-//            val end = if (index < refRangeOffsetsArray.size - 1) refRangeOffsetsArray[index + 1].toInt() else refRangeRawData.length
-//            if (offset.toInt() < end) refRangeRawData.substring(offset.toInt(), end).trimEnd('\u0000') else null
-//        }
-
-        // Extract RefRange data
-        val refRangeOffsetsArrayResult = refRangeBuffers.second.toJavaArray() as LongArray
-        val refRangeRawData = String(refRangeBuffers.first.toJavaArray() as ByteArray)
-        val refRanges = refRangeOffsetsArrayResult.take(query.resultBufferElements()["RefRange"]?.first?.toInt() ?: 0)
-            .mapIndexedNotNull { index, offset ->
-                val end = if (index < refRangeOffsetsArrayResult.size - 1) refRangeOffsetsArrayResult[index + 1].toInt()
-                else refRangeRawData.length
-                refRangeRawData.substring(offset.toInt(), end).trimEnd('\u0000')
-                if (offset.toInt() < end) {
-                    refRangeRawData.substring(offset.toInt(), end).trimEnd('\u0000')
-                } else {
-                    null
-                }
-            }
-
-        // Extract SampleName data
-        val sampleNameOffsetsArray = sampleNameOffsets.toJavaArray() as LongArray
-        val sampleNameRawData = String(sampleNameBuffer.toJavaArray() as ByteArray)
-        val sampleNames = sampleNameOffsetsArray.take(numSampleNames).mapIndexedNotNull { index, offset ->
-            val end = if (index < sampleNameOffsetsArray.size - 1) sampleNameOffsetsArray[index + 1].toInt() else sampleNameRawData.length
-            if (offset.toInt() >=0 && offset.toInt() < end && end <= sampleNameRawData.length) {
-                sampleNameRawData.substring(offset.toInt(), end).trimEnd('\u0000')
-            } else null
-        }
-
-        // Extract ID data
-        val idOffsetsArray = idOffsets.toJavaArray() as LongArray
-        val idRawData = String(idBuffer.toJavaArray() as ByteArray)
-        val ids = idOffsetsArray.take(numIDs).mapIndexedNotNull { index, offset ->
-            val end = if (index < idOffsetsArray.size - 1) idOffsetsArray[index + 1].toInt() else idRawData.length
-            if (offset.toInt() >=0 && offset.toInt() < end && end <= idRawData.length) {
-                idRawData.substring(offset.toInt(), end).trimEnd('\u0000')
-            } else null
-        }
-
-        // Combine results into a list of map
-//        val results = refRanges.zip(sampleNames.zip(ids)) { refRange, (sampleName, id) ->
-//            mapOf(
-//                "RefRange" to refRange,
-//                "SampleName" to sampleName,
-//                "ID" to id
-//            )
-//        }
-//        combinedResult.addAll(results)
-
-        // TRy this instead of above: No - doesn't help.  STill not getting the correct
-        // results.  Not sure if the problem is reading the data or storing the data.
-        refRanges.forEachIndexed { index, refRange ->
-            val sampleName = sampleNames.getOrNull(index) ?: ""
-            val id = ids.getOrNull(index) ?: ""
-            resultMap[refRange] = resultMap.getOrDefault(refRange,emptyList()) + mapOf("SampleName" to sampleName, "ID" to id)
-        }
-
-    } while (query.queryStatus == QueryStatus.TILEDB_INCOMPLETE)
-
-    // Close resources
-    query.close()
-    array.close()
-    context.close()
-    // Above got ALL The ref ranges.  I don't know why since we gave it a buffer
-    // with just a couple.  So now filter the results to just what we requested
-    //val filteredResults = combinedResult.filter { it["RefRange"] in refRangeList }
-    val filteredResults = resultMap.filterKeys { it in refRangeList }
-    println("queryWithStreaming: number of  resultMap entries = ${resultMap.keys.size}, num filtered results = ${filteredResults.keys.size}")
-    println("\nHere are all the values from the resultMap before filtering:")
-    resultMap.forEach { (key, value) -> println("Key: $key, Value: $value") }
-    //println("\nHere are all the values from the original combinedResult map before filtering:")
-    // print the values in combinedResult.  These turn out to be the same as the resultMap
-    //combinedResult.forEach { println(it) }
-
-    return filteredResults
-}
-
-
-/**
- * Given a list of reference ranges, return the sampleNames and Ids (hapids) associated with each range.
- *
- */
-fun querySampleNamesAndIDsByRefRange(arrayName: String, refRangeList: List<String>): Map<String, List<Map<String, String>>> {
-    val context = Context()
-    val result = mutableMapOf<String, List<Map<String, String>>>()
-
-    // Open the array in read mode
-    val array = Array(context, arrayName, QueryType.TILEDB_READ)
-
-    // Prepare buffers for RefRange
-    val refRangeBuffers = prepareVariableBuffer(context, refRangeList)
-    println("querySampleNamesAndIDsByRefRange: RefRangeBuffers.first = ${refRangeBuffers.first}")
-    println("querySampleNamesAndIDsByRefRange: RefRangeBuffers.second = ${refRangeBuffers.second}")
-
-    // Prepare buffers for SampleName and ID
-    val sampleNameBuffer = io.tiledb.java.api.NativeArray(context, 4096, Datatype.TILEDB_STRING_ASCII)
-    val sampleNameOffsets = io.tiledb.java.api.NativeArray(context, 512, Datatype.TILEDB_UINT64)
-
-    val idBuffer = io.tiledb.java.api.NativeArray(context, 4096, Datatype.TILEDB_STRING_ASCII)
-    val idOffsets = io.tiledb.java.api.NativeArray(context, 512, Datatype.TILEDB_UINT64)
-
-    // Create and configure the query
-    val query = Query(array, QueryType.TILEDB_READ).apply {
-        setLayout(Layout.TILEDB_GLOBAL_ORDER)
-
-        // Set buffers for RefRange (dimension filter)
-        setDataBuffer("RefRange", refRangeBuffers.first)
-        setOffsetsBuffer("RefRange", refRangeBuffers.second)
-
-        // Set buffers for SampleName (dimension retrieval)
-        setDataBuffer("SampleName", sampleNameBuffer)
-        setOffsetsBuffer("SampleName", sampleNameOffsets)
-
-        // Set buffers for ID (attribute retrieval)
-        setDataBuffer("ID", idBuffer)
-        setOffsetsBuffer("ID", idOffsets)
-    }
-
-    // Submit the query
-    query.submit()
-
-    // Use resultBufferElements to determine the number of valid entries
-    val numSampleNames = query.resultBufferElements()["SampleName"]?.first?.toInt() ?: 0
-    val numIDs = query.resultBufferElements()["ID"]?.first?.toInt() ?: 0
-    val numRefRanges = query.resultBufferElements()["RefRange"]?.first?.toInt() ?: 0
-    println("LCJ - Num Results: SampleNames=$numSampleNames, IDs=$numIDs, RefRanges=$numRefRanges")
-
-    // Extract SampleName data
-    val sampleNameOffsetsArray = sampleNameOffsets.toJavaArray() as LongArray
-    val sampleNameRawData = String(sampleNameBuffer.toJavaArray() as ByteArray)
-    println("Offsets: ${sampleNameOffsetsArray.contentToString()}")
-    println("Raw Data: $sampleNameRawData")
-    println("Num SampleNames: $numSampleNames")
-
-    val sampleNames = sampleNameOffsetsArray.take(numSampleNames).mapIndexedNotNull { index, offset ->
-        val end = if (index < numSampleNames - 1) {
-            sampleNameOffsetsArray[index + 1].toInt() // Next offset as end
-        } else {
-            sampleNameRawData.length // Last string ends at the end of raw data
-        }
-
-        if (offset.toInt() >= 0 && offset.toInt() < end && end <= sampleNameRawData.length) {
-            sampleNameRawData.substring(offset.toInt(), end).trimEnd('\u0000')
-        } else {
-            null // Skip invalid offsets
-        }
-    }
-    println("Parsed SampleNames: $sampleNames")
-
-    // Extract ID data
-    val idOffsetsArray = idOffsets.toJavaArray() as LongArray
-    val idRawData = String(idBuffer.toJavaArray() as ByteArray)
-
-    val ids = idOffsetsArray.take(numIDs).mapIndexedNotNull { index, offset ->
-        val end = if (index < numIDs - 1) {
-            idOffsetsArray[index + 1].toInt() // Next offset as end
-        } else {
-            idRawData.length // Last string ends at the end of raw data
-        }
-
-        if (offset.toInt() >= 0 && offset.toInt() < end && end <= idRawData.length) {
-            idRawData.substring(offset.toInt(), end).trimEnd('\u0000')
-        } else {
-            null // Skip invalid offsets
-        }
-    }
-    println("Parsed IDs: $ids")
-
-    // Combine results grouped by RefRange
-    val refRangeOffsetsArray = refRangeBuffers.second.toJavaArray() as LongArray
-    val refRangeRawData = String(refRangeBuffers.first.toJavaArray() as ByteArray)
-
-    // LCJ - raw data is NOT coming back correctly - I'm missing 2:1-1000
-    println("RefRangeOffsets: ${refRangeOffsetsArray.contentToString()}")
-    println("RefRangeRaw Data: $refRangeRawData")
-    val refRanges = refRangeOffsetsArray.take(numRefRanges).mapIndexedNotNull { index, offset ->
-        val end = if (index < refRangeOffsetsArray.size - 1) refRangeOffsetsArray[index + 1].toInt() else refRangeRawData.length
-        if (offset.toInt() < end) refRangeRawData.substring(offset.toInt(), end).trimEnd('\u0000') else null
-    }
-
-    // This only shows 2 ref ranges:  they are both 1:1-1000, but one for LineA and one for LineB.
-    // What happened to the other ref ranges?  I sent in 1:1-1000","2:1-1000
-    println("querySampleNamesAndIDsByRefRange: Parsed RefRanges: $refRanges")
-    refRanges.forEachIndexed { index, refRange ->
-        val sampleName = sampleNames.getOrNull(index) ?: ""
-        val id = ids.getOrNull(index) ?: ""
-        result[refRange] = result.getOrDefault(refRange, emptyList()) + mapOf(
-            "SampleName" to sampleName,
-            "ID" to id
-        )
-    }
-
-    // Close resources
-    query.close()
-    array.close()
-    context.close()
-
-    return result
-}
-
-// Given a list of Reference Ranges, query the array for all IDs associated with each Reference Range
-fun queryIDsByRefRange(arrayName: String, refRangesToQuery: List<String>): Map<String, List<String>> {
-    val context = Context()
-
-    // This one is not a list.  We have a list of refRanges to query, and we want to return a map of refRanges to IDs
-    // SO the return is a Map<String, List<String>>.  The key is the refRange, and the value is a list of hapIDs that
-    // are associated with each range.
-    val result = mutableMapOf<String, List<String>>()
-
-    // Open the array in read mode
-    val array = Array(context, arrayName, QueryType.TILEDB_READ)
-
-    // Prepare buffers for reading IDs and RefRanges
-    val idBuffer = NativeArray(context, 4096, Datatype.TILEDB_STRING_ASCII) // Adjust size as needed
-    val idOffsetsBuffer = NativeArray(context, 512, Datatype.TILEDB_UINT64)
-
-    val refRangeBuffer = NativeArray(context, 4096, Datatype.TILEDB_STRING_ASCII) // Adjust size as needed
-    val refRangeOffsetsBuffer = NativeArray(context, 512, Datatype.TILEDB_UINT64)
-
-    // Create a query object
-    val query = Query(array, QueryType.TILEDB_READ).apply {
-        setLayout(Layout.TILEDB_GLOBAL_ORDER)
-
-        // Set buffers for ID and RefRange
-        setDataBuffer("ID", idBuffer)
-        setOffsetsBuffer("ID", idOffsetsBuffer)
-
-        setDataBuffer("RefRange", refRangeBuffer)
-        setOffsetsBuffer("RefRange", refRangeOffsetsBuffer)
-    }
-
-    // Submit the query
-    query.submit()
-
-    // Process results
-    val resultBufferElements = query.resultBufferElements()
-
-    // NativeArray from TileDB doesn't have a toByteArray() method and doesn't support
-    // conversion using Kotlin's toByteARra90.  So we move to Java to do the conversion
-    val idOffsetsArray = idOffsetsBuffer.toJavaArray() as LongArray
-    //println("LCJ queryIdByRefRange - idOffsetsArray: ${idOffsetsArray.contentToString()}")
-    val idRawData = String(idBuffer.toJavaArray() as ByteArray)
-
-    val ids = idOffsetsArray.mapIndexed { index, offset ->
-        val end = if (index < idOffsetsArray.size - 1) idOffsetsArray[index + 1].toInt() else idRawData.length
-        //idRawData.substring(offset.toInt(), end).trimEnd('\u0000')
-        if (offset.toInt() < end) idRawData.substring(offset.toInt(), end).trimEnd('\u0000') else ""
-    }.filter { it.isNotEmpty() } // Remove empty strings
-
-    val refRangeOffsetsArray = refRangeOffsetsBuffer.toJavaArray() as LongArray
-    val refRangeRawData = String(refRangeBuffer.toJavaArray() as ByteArray)
-//    println("LCJ - refRangeRawData: $refRangeRawData")
-//    println("LCJ - refRangeOffsetsArray: ${refRangeOffsetsArray.contentToString()}")
-    val refRanges = refRangeOffsetsArray.mapIndexed { index, offset ->
-        val end = if (index < refRangeOffsetsArray.size - 1) refRangeOffsetsArray[index + 1].toInt() else refRangeRawData.length
-        if (offset.toInt() < end) refRangeRawData.substring(offset.toInt(), end).trimEnd('\u0000') else ""
-    }.filter { it.isNotEmpty() } // Remove empty strings
-
-    // Filter IDs by RefRange
-    refRangesToQuery.forEach { refRangeQuery ->
-        val matchingIDs = ids.zip(refRanges)
-            .filter { it.second == refRangeQuery }
-            .map { it.first }
-        result[refRangeQuery] = matchingIDs
-    }
-
-    // Close resources
-    query.close()
-    array.close()
-    context.close()
-
-    return result
-}
-
