+++ src/test/kotlin/net/maizegenetics/phgv2/utils/TiledbCoreHvcfUtilsTest.kt
@@ -139,247 +139,4 @@ class TiledbCoreHvcfUtilsTest {
         println("Finished extracting data")
     }
 
-    @Test
-    fun testQuerySampleNameID_byRefRange() {
-            // testing output from parseTiledbAltHeaders
-            var vcfReader = VCFFileReader(File(lineAhvcf), false)
-            val altHeadersLineA = parseTiledbAltHeaders(vcfReader)
-            println("Finished parsing lineAhvcf")
-
-            vcfReader = VCFFileReader(File(lineBhvcf), false)
-            val altHeadersLineB = parseTiledbAltHeaders(vcfReader)
-            println("Finished parsing lineBhvcf")
-
-            // Create the tiledb array by calling the function createTileDBArray
-            println("Creating tiledb array")
-            // ensure the top folder exists
-            File(dbPath).mkdirs()
-            createTileDBCoreArrays(dbPath)
-
-            // Write the altHeaders to the tiledb array
-            println("Writing LineA altHeaders to tiledb array")
-            writeAltDataToTileDB(altHeaderArray, altHeadersLineA)
-            println("Writing LineB altHeaders to tiledb array")
-            writeAltDataToTileDB(altHeaderArray, altHeadersLineB)
-
-            // To test what was written, we must extract data
-            println("TestCase: Extracting data from tiledb array")
-
-            val rangesToQuery = listOf("1:1-1000","1:6501-11000","2:1-1000","2:5501-6500")
-
-        // Try with dynamic buffers
-        val streamingResults = queryWithStreaming_sampleNameIdByRefRange(altHeaderArray, rangesToQuery)
-        println("\nStreaming buffer results:")
-        streamingResults.forEach { println(it) }
-
-            // delete the tiledbArray so next tests can recreate what they need
-            File(dbPath).deleteRecursively()
-    }
-
-    @Test
-    fun testQueryByRefRange() {
-
-        // testing output from parseTiledbAltHeaders
-        var vcfReader = VCFFileReader(File(lineAhvcf), false)
-        val altHeadersLineA = parseTiledbAltHeaders(vcfReader)
-        println("Finished parsing LineA alt headers ")
-        vcfReader = VCFFileReader(File(lineBhvcf), false)
-        val altHeadersLineB = parseTiledbAltHeaders(vcfReader)
-        println("Finished parsing lineBhvcf alt headers")
-
-        // Create the tiledb array by calling the function createTileDBArray
-        println("Creating tiledb array")
-        //createTileDBArraySingleDimensionID(tiledbArrayName)
-        createTileDBCoreArrays(dbPath)
-
-        // Write the altHeaders to the tiledb array
-        println("Writing altHeaders to tiledb array")
-        writeAltDataToTileDB(altHeaderArray, altHeadersLineA)
-        writeAltDataToTileDB(altHeaderArray, altHeadersLineB)
-        // Try to query by refRange
-        val resultsByRefRange = queryIDsByRefRange(altHeaderArray, listOf("1:1-1000"))
-        println("Results with 1 refRange: $resultsByRefRange")
-        val resultsByRefRange2 = queryIDsByRefRange(altHeaderArray, listOf("1:1-1000","2:12001-16500"))
-        println("Results with 2 refRanges: $resultsByRefRange2")
-
-        val resultsByRefRange5Ranges = queryIDsByRefRange(altHeaderArray, listOf("1:1-1000","1:12001-16500","1:33001-34000","2:1-1000","2:5501-6500"))
-        println("Results with 5 refRanges: $resultsByRefRange5Ranges")
-
-        // Need some asserts here !!
-
-        // delete the tiledbArray so next tests can recreate what they need
-        File(dbPath).deleteRecursively()
-
-    }
-
-    @Test
-    fun testQueryByRefRange_2() {
-
-        // Trying to see why I can filter on specific refRanges and get it correct for the alt header array,
-        // but cannot get it correct for the variants array  Added the variants array processing here,
-        // which is a copy of the test above.  Wanted to be sure I didn't mess up the test above which is
-        // the reason for the duplicate test case.
-
-        // Make sure the top folder exists
-        File(dbPath).mkdirs()
-
-        // testing output from parseTiledbAltHeaders
-        var vcfReader = VCFFileReader(File(lineAhvcf), false)
-        val altHeadersLineA = parseTiledbAltHeaders(vcfReader)
-        println("Finished parsing LineA alt headers ")
-        val variantDataLineA = parseTiledbVariantData(vcfReader)
-        println("Finished parsing lineAhvcf variant data")
-
-        vcfReader = VCFFileReader(File(lineBhvcf), false)
-        val altHeadersLineB = parseTiledbAltHeaders(vcfReader)
-        println("Finished parsing lineBhvcf alt headers")
-        val variantDataLineB = parseTiledbVariantData(vcfReader)
-        println("Finished parsing lineBhvcf variant data")
-
-        // Create the tiledb array by calling the function createTileDBArray
-        println("Creating tiledb array")
-        //createTileDBArraySingleDimensionID(tiledbArrayName)
-        createTileDBCoreArrays(dbPath)
-
-        // Write the altHeaders to the tiledb array
-        println("Writing altHeaders to tiledb array")
-        writeAltDataToTileDB(altHeaderArray, altHeadersLineA)
-        writeAltDataToTileDB(altHeaderArray, altHeadersLineB)
-
-        //Write to the variants array
-        writeVariantsDataToTileDB(variantsArray, variantDataLineA)
-        writeVariantsDataToTileDB(variantsArray, variantDataLineB)
-
-        // Try to query by refRange
-        val resultsByRefRange = queryIDsByRefRange(altHeaderArray, listOf("1:1-1000"))
-        println("Results with 1 refRange: $resultsByRefRange")
-        val resultsByRefRange2 = queryIDsByRefRange(altHeaderArray, listOf("1:1-1000","2:12001-16500"))
-        println("Results with 2 refRanges: $resultsByRefRange2")
-
-        val resultsByRefRange5Ranges = queryIDsByRefRange(altHeaderArray, listOf("1:1-1000","1:12001-16500","1:33001-34000","2:1-1000","2:5501-6500"))
-        println("Results with 5 refRanges: $resultsByRefRange5Ranges")
-
-        // Need some asserts here !!
-
-        // Query the variants array
-        val results = queryByRefRanges(variantsArray, listOf("1:1-1000","1:12001-16500","1:33001-34000","2:1-1000","2:5501-6500"))
-        println("\nResults on variantsArray:")
-        results.forEach { println(it) }
-
-
-        // delete the tiledbArray so next tests can recreate what they need
-        File(dbPath).deleteRecursively()
-
-    }
-
-    @Test
-    fun testVariantArray() {
-        println("Creating tiledb array")
-        // ensure the top folder exists
-        File(dbPath).mkdirs()
-        createTileDBCoreArrays(dbPath)
-
-        // Load the imputed hvcf file
-        // loadImputedHvcfToTileDB(imputedHvcf, dbPath)
-        // testing output from parseTiledbAltHeaders
-        var vcfReader = VCFFileReader(File(imputedHvcf), false)
-        val altHeadersImputed = parseTiledbAltHeaders(vcfReader)
-        println("Finished parsing imputedHvcf alt headers ")
-        val variantData = parseTiledbVariantData(vcfReader)
-        println("Finished parsing imputedHvcf variant data")
-
-        // Now load the data
-        val arrayName = "${dbPath}/alt_header_array"
-        writeAltDataToTileDB(arrayName, altHeadersImputed)
-        val variantArrayName = "${dbPath}/hvcf_variants_array"
-        writeVariantsDataToTileDB(variantArrayName, variantData)
-
-        // Do again for LineA and LineB
-        vcfReader = VCFFileReader(File(lineAhvcf), false)
-        val altHeadersLineA = parseTiledbAltHeaders(vcfReader)
-        writeAltDataToTileDB(arrayName, altHeadersLineA)
-        val variantDataLineA = parseTiledbVariantData(vcfReader)
-        writeVariantsDataToTileDB(variantArrayName, variantDataLineA)
-        println("Finished writing LineA data")
-
-        vcfReader = VCFFileReader(File(lineBhvcf), false)
-        val altHeadersLineB = parseTiledbAltHeaders(vcfReader)
-        writeAltDataToTileDB(arrayName, altHeadersLineB)
-        val variantDataLineB = parseTiledbVariantData(vcfReader)
-        writeVariantsDataToTileDB(variantArrayName, variantDataLineB)
-
-        // Query for distinct sampleNames from the hvcf_variants_array
-        println("QUery distinct sample names from the hvcf_variants_array")
-        val names = queryDistinctSampleNames(variantArrayName)
-        println("Distinct sample names: $names")
-        assertTrue(names.size == 3)
-        assertTrue(names.contains("LineA"))
-        assertTrue(names.contains("LineB"))
-        assertTrue(names.contains("TestLine2"))
-
-        //Query for distinct refRanges from the hvcf_variants_array
-        println("Query distinct refRanges from the hvcf_variants_array")
-        val refRangeList = queryDistinctRefRanges(variantArrayName)
-        println("\nnumber of distinct ref ranges: ${refRangeList.size}")
-        println("Distinct refRanges: $refRangeList")
-        assertEquals(38, refRangeList.size)
-
-
-        // QUery for SampleName and ID based on a list of RefRanges
-        // I can't get this to work.  It is fine when I try the query on the altHeaderArray
-        //  but not on the variantsArray.
-
-        //val refRanges = listOf("1:1-1000","1:12001-16500","1:33001-34000","2:1-1000","2:5501-6500")
-        val refRanges = listOf("1:33001-34000","2:1-1000","2:5501-6500")
-        //val results = queryByRefRanges(variantArrayName, refRanges)
-        val results = queryByRefRanges(variantArrayName, listOf("1:1-1000","1:12001-16500","1:33001-34000","2:1-1000","2:5501-6500"))
-        println("\nResults on variantsArray:")
-        results.forEach { println(it) }
-
-        // Query as we do the altHeaders
-        val resultsByRefRange2 = queryIDsByRefRange(altHeaderArray, listOf("1:1-1000","1:12001-16500","1:33001-34000","2:1-1000","2:5501-6500"))
-
-        println("\nResults with 5 refRanges:, altHeaderArray: ")
-        results.forEach {println(it)}
-
-        File(dbPath).deleteRecursively()
-    }
-
-    @Test
-    fun testLoadImputedHvcf() {
-        // TODO - this failes - no output!1 Just columns, no rows.
-        // Load the imputed hvcf file into the tiledb core arrays
-        // Create the tiledb array by calling the function createTileDBArray
-        println("Creating tiledb array")
-        // ensure the top folder exists
-        File(dbPath).mkdirs()
-        createTileDBCoreArrays(dbPath)
-
-        // Load the imputed hvcf file
-       // loadImputedHvcfToTileDB(imputedHvcf, dbPath)
-        // testing output from parseTiledbAltHeaders
-        var vcfReader = VCFFileReader(File(imputedHvcf), false)
-        val altHeadersImputed = parseTiledbAltHeaders(vcfReader)
-        println("Finished parsing imputedHvcf alt headers ")
-        val variantData = parseTiledbVariantData(vcfReader)
-        println("Finished parsing imputedHvcf variant data")
-
-        // Now load the data
-        val arrayName = "${dbPath}/alt_header_array"
-        writeAltDataToTileDB(arrayName, altHeadersImputed)
-        val variantArrayName = "${dbPath}/hvcf_variants_array"
-        writeVariantsDataToTileDB(variantArrayName, variantData)
-
-        println("QUery distinct sample names from the hvcf_variants_array")
-        val names = queryDistinctSampleNames(variantArrayName)
-        println("Distinct sample names: $names")
-
-        // Query the data to verify it is loaded.
-        // QUery only the variants data:
-        val results = queryVariantsArray(variantArrayName, null, listOf("TestLine2"))
-
-        // delete the tiledbArray so next tests can recreate what they need
-        File(dbPath).deleteRecursively()
-    }
-
 }
